+++ src/Processor.hpp
@@ -19,7 +19,7 @@ class Processor {
   CoreMonitor monitor{};
 
   Processor() : cache(std::make_shared<Cache>()) { instructionStream << std::ifstream("data/test_0.data").rdbuf(); };
-  Processor(uint8_t pid, const std::ifstream& filePathName, uint8_t associativity, uint32_t numBlocks,
+  Processor(uint8_t pid, const std::ifstream& filePathName, uint32_t associativity, uint32_t numBlocks,
             uint16_t blockSize) :
       pid(pid),
       cache(std::make_shared<Cache>(pid, associativity, numBlocks, blockSize)) {
