+++ src/Runner.hpp
@@ -15,7 +15,7 @@ class Runner {
     std::string_view protocol{"MESI"};
     std::filesystem::path benchmark{"data/bodytrack"};
     uint32_t cacheSize{4096};
-    uint8_t associativity{2};
+    uint32_t associativity{2};
     uint16_t blockSize{32};
     uint32_t numBlocks{4096 / 32};
